update squirrel library

+++ src/squirrel/squirrel.cpp
@@ -89,6 +89,11 @@ squirrel::~squirrel()
 bool squirrel::DatabaseConnect() {
 
     db = QSqlDatabase::addDatabase("QSQLITE", "squirrel");
+    if (!db.isValid()) {
+        Log(QString("Error initializing SQLite database (likely driver related) [%1]. Error [%2]").arg(db.databaseName()).arg(db.lastError().text()), __FUNCTION__);
+        return false;
+    }
+
     if (debug) {
         QFile::remove("/tmp/sqlite.db");
         db.setDatabaseName("/tmp/sqlite.db");
@@ -101,7 +106,7 @@ bool squirrel::DatabaseConnect() {
         return true;
     }
     else {
-        Log(QString("Error opening SQLite memory database [%1]. Error [%2]").arg(db.databaseName()).arg(db.lastError().text()), __FUNCTION__);
+        Log(QString("Error opening SQLite database [%1]. Error [%2]").arg(db.databaseName()).arg(db.lastError().text()), __FUNCTION__);
         return false;
     }
 }
@@ -844,6 +849,18 @@ bool squirrel::Validate() {
 
 
 /* ------------------------------------------------------------ */
+/* ----- GetLogBuffer ----------------------------------------- */
+/* ------------------------------------------------------------ */
+QString squirrel::GetLogBuffer() {
+
+    QString ret = logBuffer;
+    logBuffer = "";
+
+    return ret;
+}
+
+
+/* ------------------------------------------------------------ */
 /* ----- Print ------------------------------------------------ */
 /* ------------------------------------------------------------ */
 /**
@@ -1218,6 +1235,7 @@ void squirrel::Log(QString s, QString func) {
     if (!quiet) {
         if (s.trimmed() != "") {
             log.append(QString("squirrel::%1() %2\n").arg(func).arg(s));
+            logBuffer.append(QString("squirrel::%1() %2\n").arg(func).arg(s));
             utils::Print(QString("squirrel::%1() %2").arg(func).arg(s));
         }
     }
@@ -1231,6 +1249,7 @@ void squirrel::Debug(QString s, QString func) {
     if (debug) {
         if (s.trimmed() != "") {
             log.append(QString("Debug squirrel::%1() %2\n").arg(func).arg(s));
+            logBuffer.append(QString("Debug squirrel::%1() %2\n").arg(func).arg(s));
             utils::Print(QString("Debug squirrel::%1() %2").arg(func).arg(s));
         }
     }
